+++ examples/max-depth.js
@@ -1,3 +1,8 @@
+// This file is distributed under the MIT license.
+// See LICENSE file in the project root for details.
+
+'use strict';
+
 function getDepth(element, depth) {
   let maxDepth = depth;
   for (let i = 0; i < element.children.length; ++i) {