+++ examples/navigation.js
@@ -1,6 +1,3 @@
-//
-// Copyright (c) 2017 Masayuki Nagamachi <masayuki.nagamachi@gmail.com>
-//
 // This file is distributed under the MIT license.
 // See LICENSE file in the project root for details.