+++ examples/tags.js
@@ -1,3 +1,8 @@
+// This file is distributed under the MIT license.
+// See LICENSE file in the project root for details.
+
+'use strict';
+
 let elements = document.querySelectorAll('*');
 let tags = {};
 for (let i = 0; i < elements.length; ++i) {