+++ test/navigation.js
@@ -1,7 +1,6 @@
-//
-// Copyright (c) 2017 Masayuki Nagamachi <masayuki.nagamachi@gmail.com>
-//
 // This file is distributed under the MIT license.
 // See LICENSE file in the project root for details.
+
+'use strict';
 
 module.exports.navigate = (driver) => driver;